+++ plugins/variants/src/VariantFeatureWidget/LaunchBreakendPanel/LaunchBreakendPanel.tsx
@@ -4,7 +4,7 @@ import BaseCard from '@jbrowse/core/BaseFeatureWidget/BaseFeatureDetail/BaseCard
 import { SimpleFeature, getEnv, getSession } from '@jbrowse/core/util'
 import { Link, Typography } from '@mui/material'
 
-import type { VariantFeatureWidgetModel } from './stateModelFactory'
+import type { VariantFeatureWidgetModel } from '../stateModelFactory'
 import type { SimpleFeatureSerialized } from '@jbrowse/core/util'
 
 // lazies
