+++ plugins/variants/src/VariantFeatureWidget/LaunchBreakendPanel/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './LaunchBreakendPanel'
