+++ plugins/variants/src/VariantFeatureWidget/VariantSampleGrid/VariantSampleGrid.tsx
