+++ plugins/variants/src/VariantFeatureWidget/VariantSampleGrid/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './VariantSampleGrid'
